Added more constant values to MathCraft, split moddedblocks and vanillablock overrides into separate classes
