+++ src/main/java/com/github/torisen13/dirtfall/dirtfall/setup/ModBlocks.java
@@ -47,14 +47,6 @@ public class ModBlocks {
                     .sound(SoundType.METAL)),
             ItemGroup.TAB_MISC);
 
-    // Attempting to override a vanilla block
-    public static final RegistryObject<Block> DIRT = registerOverrideBlockWithBlockItem("dirt",
-            () -> new FallingBlock(AbstractBlock.Properties
-                    .of(Material.DIRT, MaterialColor.DIRT)
-                    .strength(0.5F)
-                    .sound(SoundType.GLASS)),
-            ItemGroup.TAB_BUILDING_BLOCKS);
-
     static void register() {}
 
     // Helper method to register a Block
@@ -67,19 +59,6 @@ public class ModBlocks {
     private static <T extends Block> RegistryObject<T> registerBlockWithBlockItem(String name, Supplier<T> block, ItemGroup itemGroup) {
         RegistryObject<T> newBlock = registerBlock(name, block);
         Registration.ITEMS.register(name, () -> new BlockItem(newBlock.get(), new Item.Properties().tab(itemGroup)));
-        return newBlock;
-    }
-
-    // Helper method to register a Block override
-    private static <T extends Block> RegistryObject<T> registerOverrideBlock(String name, Supplier<T> block) {
-        RegistryObject<T> newBlock = Registration.OVERRIDE_BLOCKS.register(name, block);
-        return newBlock;
-    }
-
-    // Helper method to register an override Block with a corresponding BlockItem
-    private static <T extends Block> RegistryObject<T> registerOverrideBlockWithBlockItem(String name, Supplier<T> block, ItemGroup itemGroup) {
-        RegistryObject<T> newBlock = registerOverrideBlock(name, block);
-        Registration.OVERRIDE_ITEMS.register(name, () -> new BlockItem(newBlock.get(), new Item.Properties().tab(itemGroup)));
         return newBlock;
     }
 }
